[Added] Encoding dialog. Set/GetEncoding public methods. [Changes] Rename HEXBOOKMARKSTRUCT->HEXBKMSTRUCT. Lots of internal cleanups, improvements and simplifications.

+++ HexCtrl/src/CScrollEx.cpp
@@ -16,7 +16,7 @@ using namespace HEXCTRL::INTERNAL::SCROLLEX;
 
 namespace HEXCTRL::INTERNAL::SCROLLEX
 {
-	enum class EState : DWORD
+	enum class EState : WORD
 	{
 		STATE_DEFAULT,
 		FIRSTARROW_HOVER, FIRSTARROW_CLICK,
